requirements.txt
# Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
supabase==2.0.3
requests==2.31.0

# Frontend
streamlit==1.28.2
streamlit-autorefresh==1.0.1
plotly==5.18.0
pandas==2.1.3
folium==0.15.1
streamlit-folium==0.15.1

# Utilities
python-dateutil==2.8.2
polyline==2.0.0
